SyncServer adding functionality to Loader
In one big commit as PR wasnt merged before rebranding and merge exploded
