+++ openpype/modules/sync_server/providers/abstract_provider.py
@@ -1,16 +1,22 @@
-from abc import ABCMeta, abstractmethod
+import abc, six
+from openpype.api import Anatomy, Logger
+
+log = Logger().get_logger("SyncServer")
 
 
-class AbstractProvider(metaclass=ABCMeta):
+@six.add_metaclass(abc.ABCMeta)
+class AbstractProvider:
 
-    def __init__(self, site_name, tree=None, presets=None):
+    def __init__(self, project_name, site_name, tree=None, presets=None):
         self.presets = None
         self.active = False
         self.site_name = site_name
 
         self.presets = presets
 
-    @abstractmethod
+        super(AbstractProvider, self).__init__()
+
+    @abc.abstractmethod
     def is_active(self):
         """
             Returns True if provider is activated, eg. has working credentials.
@@ -18,36 +24,54 @@ class AbstractProvider(metaclass=ABCMeta):
             (boolean)
         """
 
-    @abstractmethod
-    def upload_file(self, source_path, target_path, overwrite=True):
+    @abc.abstractmethod
+    def upload_file(self, source_path, path,
+                    server, collection, file, representation, site,
+                    overwrite=False):
         """
             Copy file from 'source_path' to 'target_path' on provider.
             Use 'overwrite' boolean to rewrite existing file on provider
 
         Args:
-            source_path (string): absolute path on local system
-            target_path (string): absolute path on provider (GDrive etc.)
-            overwrite (boolean): True if overwite existing
+            source_path (string):
+            path (string): absolute path with or without name of the file
+            overwrite (boolean): replace existing file
+
+            arguments for saving progress:
+            server (SyncServer): server instance to call update_db on
+            collection (str): name of collection
+            file (dict): info about uploaded file (matches structure from db)
+            representation (dict): complete repre containing 'file'
+            site (str): site name
         Returns:
             (string) file_id of created file, raises exception
         """
         pass
 
-    @abstractmethod
-    def download_file(self, source_path, local_path, overwrite=True):
+    @abc.abstractmethod
+    def download_file(self, source_path, local_path,
+                      server, collection, file, representation, site,
+                      overwrite=False):
         """
             Download file from provider into local system
 
         Args:
             source_path (string): absolute path on provider
-            local_path (string): absolute path on local
-            overwrite (bool): default set to True
+            local_path (string): absolute path with or without name of the file
+            overwrite (boolean): replace existing file
+
+            arguments for saving progress:
+            server (SyncServer): server instance to call update_db on
+            collection (str): name of collection
+            file (dict): info about uploaded file (matches structure from db)
+            representation (dict): complete repre containing 'file'
+            site (str): site name
         Returns:
             None
         """
         pass
 
-    @abstractmethod
+    @abc.abstractmethod
     def delete_file(self, path):
         """
             Deletes file from 'path'. Expects path to specific file.
@@ -60,7 +84,7 @@ class AbstractProvider(metaclass=ABCMeta):
         """
         pass
 
-    @abstractmethod
+    @abc.abstractmethod
     def list_folder(self, folder_path):
         """
             List all files and subfolders of particular path non-recursively.
@@ -72,7 +96,7 @@ class AbstractProvider(metaclass=ABCMeta):
         """
         pass
 
-    @abstractmethod
+    @abc.abstractmethod
     def create_folder(self, folder_path):
         """
             Create all nonexistent folders and subfolders in 'path'.
@@ -85,7 +109,7 @@ class AbstractProvider(metaclass=ABCMeta):
         """
         pass
 
-    @abstractmethod
+    @abc.abstractmethod
     def get_tree(self):
         """
             Creates folder structure for providers which do not provide
@@ -94,16 +118,49 @@ class AbstractProvider(metaclass=ABCMeta):
         """
         pass
 
-    @abstractmethod
-    def resolve_path(self, path, root_config, anatomy=None):
+    @abc.abstractmethod
+    def get_roots_config(self, anatomy=None):
         """
-            Replaces root placeholders with appropriate real value from
-            'root_configs' (from Settings or Local Settings) or Anatomy
-            (mainly for 'studio' site)
+            Returns root values for path resolving
 
-            Args:
-                path(string): path with '{root[work]}/...'
-                root_config(dict): from Settings or Local Settings
-                anatomy (Anatomy): prepared anatomy object for project
+            Takes value from Anatomy which takes values from Settings
+            overridden by Local Settings
+
+        Returns:
+            (dict) - {"root": {"root": "/My Drive"}}
+                     OR
+                     {"root": {"root_ONE": "value", "root_TWO":"value}}
+            Format is importing for usage of python's format ** approach
         """
         pass
+
+    def resolve_path(self, path, root_config=None, anatomy=None):
+        """
+            Replaces all root placeholders with proper values
+
+            Args:
+                path(string): root[work]/folder...
+                root_config (dict): {'work': "c:/..."...}
+                anatomy (Anatomy): object of Anatomy
+            Returns:
+                (string): proper url
+        """
+        if root_config and not root_config.get("root"):
+            root_config = {"root": root_config}
+        else:
+            root_config = self.get_roots_config(anatomy)
+
+        try:
+            if not root_config:
+                raise KeyError
+
+            path = path.format(**root_config)
+        except KeyError:
+            try:
+                path = anatomy.fill_root(path)
+            except KeyError:
+                msg = "Error in resolving local root from anatomy"
+                log.error(msg)
+                raise ValueError(msg)
+
+        return path
