+++ openpype/modules/sync_server/utils.py
@@ -1,6 +1,12 @@
 import time
-from openpype.api import  Logger
+from openpype.api import Logger
 log = Logger().get_logger("SyncServer")
+
+
+class SyncStatus:
+    DO_NOTHING = 0
+    DO_UPLOAD = 1
+    DO_DOWNLOAD = 2
 
 
 def time_function(method):
